% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_CalculateSexQCDataFrame}
\alias{array_CalculateSexQCDataFrame}
\title{Generate Sex dependent QC DataFrame.}
\usage{
array_CalculateSexQCDataFrame(study_data, assay_name)
}
\arguments{
\item{study_data}{SummarizedExperiment file to be analysed}

\item{assay_name}{Assay name to be analysed (exprs, norm_exprs or batch_exprs)}
}
\value{
DataFrame object of SEX QC
}
\description{
Generate Sex dependent QC DataFrame.
}
\author{
Nurlan Kerimov, Liis Kolberg
}
