% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_normaliseSE}
\alias{array_normaliseSE}
\title{Normalise microarray data using lumiN function}
\usage{
array_normaliseSE(se, norm_method = "quantile", assay_name = "exprs",
  log_transform = TRUE, adjust_batch = FALSE, filter_quality = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment file to be normalised}

\item{norm_method}{Normalisation method from quantile, rsn, ssn, loess, vsn, rankinvariant (Default:quantile)}

\item{assay_name}{Assay name to be normalised (exprs or batch_exprs)}

\item{log_transform}{Boolean value if intensities should be log2 transformed (Default:TRUE)}

\item{adjust_batch}{Boolean value if data should be adjusted for batch effects before normalisation (Default:FALSE)}

\item{filter_quality}{Boolean value if should exclude samples with RNA_QC_passed=FALSE (Default:TRUE)}
}
\value{
SummarizedExperiment with assay 'norm_exprs'
}
\description{
Normalise microarray data using lumiN function
}
\author{
Liis Kolberg
}
