% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_PlotMDS}
\alias{array_PlotMDS}
\title{Generate MDS QC Plot}
\usage{
array_PlotMDS(se, assay_name = "exprs", condition = NULL,
  filter_quality = TRUE, html_output = FALSE, output_dir = "./")
}
\arguments{
\item{se}{SummarizedExperiment file to be analysed}

\item{assay_name}{Assay name to be analysed (exprs, norm_exprs or batch_exprs)}

\item{condition}{List of conditions to plot, e.g naive (Default:NULL)}

\item{filter_quality}{Boolean value if samples with RNA_QC_passed = FALSE should be excluded (Default:TRUE)}

\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
MDS plot of study
}
\description{
Generate MDS QC Plot
}
\author{
Liis Kolberg
}
