% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_plotPCA}
\alias{array_plotPCA}
\title{Generate PCA QC Plot}
\usage{
array_plotPCA(se, assay_name = "exprs", filter_quality = TRUE,
  show_quality = FALSE, html_output = FALSE, output_dir = "./")
}
\arguments{
\item{se}{SummarizedExperiment file to be analysed}

\item{assay_name}{Assay name to be analysed (exprs, norm_exprs or batch_exprs)}

\item{filter_quality}{Boolean value if samples with RNA_QC_passed = FALSE should be excluded (Default:TRUE)}

\item{show_quality}{Boolean value if samples with RNA_QC_passed = FALSE should be highlighted (Default:FALSE)}

\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
PCA plot of study
}
\description{
Generate PCA QC Plot
}
\author{
Liis Kolberg
}
