% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_PlotSexQC}
\alias{array_PlotSexQC}
\title{Generate Sex dependent QC Plot.}
\usage{
array_PlotSexQC(study_data, assay_name = "exprs", html_output = FALSE,
  output_dir = "./")
}
\arguments{
\item{study_data}{SummarizedExperiment file to be analysed}

\item{assay_name}{Assay name to be analysed (exprs, norm_exprs or batch_exprs)}

\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
plotly object of SEX QC
}
\description{
Generate Sex dependent QC Plot.
}
\author{
Nurlan Kerimov, Liis Kolberg
}
