% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{array_RemoveBatch}
\alias{array_RemoveBatch}
\title{Adjust for batch effects using limma removeBatchEffect}
\usage{
array_RemoveBatch(se, assay_name = "exprs")
}
\arguments{
\item{se}{SummarizedExperiment file to be analysed}

\item{assay_name}{Assay name to be analysed (exprs, norm_exprs or batch_exprs)}
}
\value{
Batch effect adjusted SE with new assay called batch_exprs
}
\description{
Adjust for batch effects using limma removeBatchEffect
}
\author{
Liis Kolberg
}
