% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_assign.R
\name{assign_populations}
\alias{assign_populations}
\title{Assigns populations using distance matrix from generate_distance_matrix() function}
\usage{
assign_populations(distance_matrix, abs_threshold = 0.02,
  rel_threshold = 1.7)
}
\arguments{
\item{distance_matrix}{distance matrix dataframe created by generate_distance_matrix() function}

\item{abs_threshold}{sample's minimum distance to any reference population should be less than this threshold to be assigned to population. Default value is 0.02}

\item{rel_threshold}{sample's minimum distance to any reference population should be rel_threshold times smaller than second minimum distanceto be assigned to population. Default value is 1.7}
}
\description{
Assigns populations using distance matrix from generate_distance_matrix() function
}
\author{
Nurlan Kerimov
}
