% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{calculate_rna_qc_matrices}
\alias{calculate_rna_qc_matrices}
\title{Export RNA QC (PCA, MDS, Sex) matrices}
\usage{
calculate_rna_qc_matrices(rds_files_path, output_path)
}
\arguments{
\item{rds_files_path}{Path where SummarizedExperiment files live}

\item{output_path}{Path where the plots will ve saved}
}
\description{
Export RNA QC (PCA, MDS, Sex) matrices
}
\author{
Nurlan Kerimov
}
