% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{calculateMDSMatrix}
\alias{calculateMDSMatrix}
\title{Generate MDS Matrix for SummarizedExperiment}
\usage{
calculateMDSMatrix(study_data_se, condition = "all")
}
\arguments{
\item{study_data_se}{SummarizedExperiment file to be analysed}

\item{condition}{Boolean value if html output should be created (Default:FALSE)}
}
\value{
MDS Matrix of study
}
\description{
Generate MDS Matrix for SummarizedExperiment
}
\author{
Nurlan Kerimov
}
