% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{calculateSexQCDataFrame}
\alias{calculateSexQCDataFrame}
\title{Generate Sex dependent QC DataFrame.}
\usage{
calculateSexQCDataFrame(study_data)
}
\arguments{
\item{study_data}{SummarizedExperiment file to be analysed}
}
\value{
DataFrame object of SEX QC
}
\description{
Generate Sex dependent QC DataFrame.
}
\author{
Nurlan Kerimov
}
