% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{calculateTranscriptUsage}
\alias{calculateTranscriptUsage}
\title{Calculate the transcript usage}
\usage{
calculateTranscriptUsage(expression_matrix, phenotype_map)
}
\arguments{
\item{expression_matrix}{expression_matrix}

\item{phenotype_map}{phenotype_map}
}
\description{
Calculate the transcript usage
}
