% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{convertSEtoQTLtools}
\alias{convertSEtoQTLtools}
\title{Convert SummarizedExperiment object into a bed file suitable for QTLTools}
\usage{
convertSEtoQTLtools(se, assay_name = "cqn")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_name}{Assay of the SummarizedExperiment object used for QTL mapping}
}
\value{
Assay converted into bed format suitable for QTLTools
}
\description{
Convert SummarizedExperiment object into a bed file suitable for QTLTools
}
