% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_dose_matrix.R
\name{extractGenotypeMatrixFromDosage}
\alias{extractGenotypeMatrixFromDosage}
\title{Extract genotype matrix from a dosage file generated by bcftools}
\usage{
extractGenotypeMatrixFromDosage(chr, start, end, dosage_file)
}
\arguments{
\item{chr}{chromosome of the genomic region}

\item{start}{start coordinate of the genomic region}

\item{end}{end coordinate of the genomic region}

\item{dosage_file}{Path to the tabix-indexed dosage file (first four columns need to be CHROM, POS, REF, ALT)}
}
\value{
genotype matrix
}
\description{
Extract genotype matrix from a dosage file generated by bcftools
}
