% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray_helpers.R
\name{extractGenotypeMatrixFromGDS}
\alias{extractGenotypeMatrixFromGDS}
\title{Extract a genotype matrix corresponding to a specific genomic region}
\usage{
extractGenotypeMatrixFromGDS(chr, start, end, variant_information, gdsfile)
}
\arguments{
\item{chr}{Chromosome of the region}

\item{start}{Start coordinate of the region}

\item{end}{End coordinate of the reion}

\item{variant_information}{Variant infromation constructed by importVariantInformationFromGDS}

\item{gdsfile}{Path to the GDS file}
}
\value{
genotype matrix
}
\description{
Extract a genotype matrix corresponding to a specific genomic region
}
