% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{extractPhentypeFromSE}
\alias{extractPhentypeFromSE}
\title{Extract a single phenotype from a SummarizedExperiment object.}
\usage{
extractPhentypeFromSE(phenotype_id, se, assay)
}
\arguments{
\item{phenotype_id}{Id of the phenotype to be extracted}

\item{se}{SummarizedExperiment object}

\item{assay}{Name of the assay in the SummarizedExperiment object.}
}
\value{
Data frame with phenotype values and sample metadata.
}
\description{
Produces a single data frame where phenotype values have been joined into a single data frame with sample metadata.
}
