% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray_helpers.R
\name{extractVariantGenotypeFromGDS}
\alias{extractVariantGenotypeFromGDS}
\title{Extract genotypes of a single genetic variant from the GDS file}
\usage{
extractVariantGenotypeFromGDS(variant_id, variant_information, gdsfile)
}
\arguments{
\item{variant_id}{Id of the genetic variant}

\item{variant_information}{Variant infromation constructed by importVariantInformationFromGDS}

\item{gdsfile}{Path to the GDS file}
}
\value{
data frame with the genotypes
}
\description{
Extract genotypes of a single genetic variant from the GDS file
}
