% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{filterSummarizedExperiment}
\alias{filterSummarizedExperiment}
\title{Filter a Summarized Experiment according to provided parameters}
\usage{
filterSummarizedExperiment(se, valid_chromosomes = NA,
  valid_gene_types = NA, filter_rna_qc = FALSE,
  filter_genotype_qc = FALSE)
}
\arguments{
\item{se}{Summarized Experiment object to be filteres}

\item{valid_chromosomes}{list (string vector) of valid gene chromosomes}

\item{valid_gene_types}{list (string vector) of valid gene types}

\item{filter_rna_qc}{Boolean value}

\item{filter_genotype_qc}{Boolean value}
}
\description{
Filter a Summarized Experiment according to provided parameters
}
\author{
Kaur Alasoo
}
