% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_assign.R
\name{generate_distance_matrix}
\alias{generate_distance_matrix}
\title{Generate distance matrix of genotype samples using PC matrices}
\usage{
generate_distance_matrix(
  reference_pca_df,
  pca_df,
  n_pcs = 3,
  method = "euclidean"
)
}
\arguments{
\item{reference_pca_df}{PC matrix of reference genotype data. Should have genotype_id, superpopulation_code and PCN columns}

\item{pca_df}{PC matrix of input genotype data. Should have genotype_id, superpopulation_code and PCN columns}

\item{n_pcs}{number of PCs to be taken into account. Default value is 3}

\item{method}{Method of distance measure. See stats::dist() method for options}
}
\description{
Generate distance matrix of genotype samples using PC matrices
}
\author{
Nurlan Kerimov
}
