% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importBiomartMetadata}
\alias{importBiomartMetadata}
\title{Import transcript metadata from biomart web export}
\usage{
importBiomartMetadata(biomart_path,
  col_types = "ccccccciciiciiiiccccccccidccccii")
}
\arguments{
\item{biomart_path}{Path to the biomart download text file}

\item{col_types}{Column types of the biomart download file (for readr)}
}
\value{
tibble containing transcript metadata
}
\description{
Import transcript metadata from biomart web export
}
