% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{importQTLtoolsPermRes}
\alias{importQTLtoolsPermRes}
\title{Import QTLtools output table from permutation run v2 into R.}
\usage{
importQTLtoolsPermRes(file_path)
}
\arguments{
\item{file_path}{Path to the QTLtools permutation output file.}
}
\value{
data_frame containing gene_ids, snp ids and p-values.
}
\description{
Import QTLtools output table from permutation run v2 into R.
}
\author{
Nurlan Kerimov
}
