% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{importQTLtoolsTable}
\alias{importQTLtoolsTable}
\title{Import QTLtools output table from permutation run into R.}
\usage{
importQTLtoolsTable(file_path)
}
\arguments{
\item{file_path}{Path to the QTLtools output file.}
}
\value{
data_frame containing gene_ids, snp ids and p-values.
}
\description{
Import QTLtools output table from permutation run into R.
}
\author{
Kaur Alasoo
}
