% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importTxreviseCounts}
\alias{importTxreviseCounts}
\title{Read and merge the TxRevise count matrices}
\usage{
importTxreviseCounts(path)
}
\arguments{
\item{path}{Path to directory where TxRevise files located.}
}
\description{
Read and merge the TxRevise count matrices
}
