% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{importVariantInformation}
\alias{importVariantInformation}
\title{Import variant information extracted from VCF file into R}
\usage{
importVariantInformation(path)
}
\arguments{
\item{path}{Path to the the variant information text file.}
}
\description{
The variant information text file can be generated from the VCF using the following
bcftools command:
bcftools query -f '\%CHROM\\t\%POS\\t\%ID\\t\%REF\\t\%ALT\\t\%TYPE\\t\%AC\\t\%AN\\n' path/to/vcf_file.vcf.gz | bgzip > path/to/variant_infromation_file.txt.gz
}
