% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDSArray_helpers.R
\name{importVariantInformationFromGDS}
\alias{importVariantInformationFromGDS}
\title{Import variant information from GDS file}
\usage{
importVariantInformationFromGDS(gdsfile)
}
\arguments{
\item{gdsfile}{Path to the genotype file in GDS format.}
}
\value{
data frame with variant information
}
\description{
Import variant information from GDS file
}
