% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_leafcutter.R
\name{leafcutterAnnotateIntrons}
\alias{leafcutterAnnotateIntrons}
\title{Annotate leafcutter introns with gene metadata}
\usage{
leafcutterAnnotateIntrons(leafcutter_ids, intron_annotation_path,
  transcript_meta)
}
\arguments{
\item{leafcutter_ids}{vector of leafcutter intron ids}

\item{intron_annotation_path}{Path to leafcutter intron annotations file}

\item{transcript_meta}{transcript metadata data frame from importBiomartMetadata() function}
}
\value{
data frame of intron metadata
}
\description{
Annotate leafcutter introns with gene metadata
}
