% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_qc_utils.R
\name{log2_transform}
\alias{log2_transform}
\title{Log2 transform intensity values (same approach as in function lumiB)}
\usage{
log2_transform(mat)
}
\arguments{
\item{mat}{Matrix of intensity values}
}
\value{
Log2 transformed matrix
}
\description{
Log2 transform intensity values (same approach as in function lumiB)
}
\author{
Liis Kolberg
}
