% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{makeSummarizedExperimentFromCountMatrix}
\alias{makeSummarizedExperimentFromCountMatrix}
\title{Make Summarized Experiment object from raw count matrix}
\usage{
makeSummarizedExperimentFromCountMatrix(assay, row_data, col_data,
  assay_name = "counts", quant_method = "gene_counts")
}
\arguments{
\item{assay}{count matrix of any phenotype quantification.}

\item{row_data}{phenotype metadata}

\item{col_data}{sample metadata}

\item{assay_name}{custom assay name to be created in SE. Default "counts"}

\item{quant_method}{Quantification method. Can be gene_counts, leafcutter, txrevise, transcript_usage or exon_counts}
}
\description{
Make Summarized Experiment object from raw count matrix
}
\author{
Nurlan Kerimov
}
