% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{mbvFindBestMatch}
\alias{mbvFindBestMatch}
\title{Post-process QTLTools mbv results to find the best matching individual for each sample}
\usage{
mbvFindBestMatch(mbv_df)
}
\arguments{
\item{mbv_df}{Data frame with MBV results for one sequencing sample.}
}
\value{
Data frame with one row identifying the best matching invidivual for this sample
}
\description{
Post-process QTLTools mbv results to find the best matching individual for each sample
}
