% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{mbvImportData}
\alias{mbvImportData}
\title{Import MBV output files for all samples in a studt}
\usage{
mbvImportData(mbv_dir, suffix = ".mbv_output.txt")
}
\arguments{
\item{mbv_dir}{directory where mbv output files are located}
}
\value{
List of the matrixes with mbv output values per sample
}
\description{
Import MBV output files for all samples in a studt
}
