% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{normaliseSE_cqn}
\alias{normaliseSE_cqn}
\title{Normalises the assay of SE with cqn method}
\usage{
normaliseSE_cqn(se, assay_name = "counts")
}
\arguments{
\item{se}{Summarized Experiment object to}

\item{assay_name}{custom assay_name (default: counts)}
}
\description{
Normalises the assay of SE with cqn method
}
\author{
Kaur Alasoo
}
