% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{normaliseSE_quantile}
\alias{normaliseSE_quantile}
\title{Quantile normalise SummarizedExperiment by rows}
\usage{
normaliseSE_quantile(se, assay_name = "usage")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_name}{Name of the assay to be normalised in the se object}
}
\value{
SummarizedExperiment object with quantile-normalised data in the qnorm assay
}
\description{
Quantile normalise SummarizedExperiment by rows
}
