% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{normaliseSE_ratios}
\alias{normaliseSE_ratios}
\title{normaliseSE_ratios}
\usage{
normaliseSE_ratios(se, assay_name = "tpms")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_name}{custom assay name for generated assay}
}
\value{
SummarizedExperiment object with additional assay
}
\description{
normaliseSE_ratios
}
