% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{normaliseSE_tpm}
\alias{normaliseSE_tpm}
\title{Normalises the assay of SE with TPM method}
\usage{
normaliseSE_tpm(se, assay_name = "counts", fragment_length = 250)
}
\arguments{
\item{se}{Summarized Experiment object}

\item{assay_name}{custom assay_name (default: counts)}

\item{fragment_length}{custom fragment_length (default: 250)}
}
\description{
Normalises the assay of SE with TPM method
}
\author{
Kaur Alasoo
}
