% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plot_mbv_results}
\alias{plot_mbv_results}
\title{Generate MBV_results plots}
\usage{
plot_mbv_results(mbv_files_path, output_path)
}
\arguments{
\item{mbv_files_path}{Path where mbv_output files live}

\item{output_path}{Path where the plots will ve saved}
}
\description{
Generate MBV_results plots
}
\author{
Nurlan Kerimov
}
