% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plot_rna_qc_all}
\alias{plot_rna_qc_all}
\title{Generate RNA QC (PCA, MDS, Sex) plots}
\usage{
plot_rna_qc_all(rds_files_path, output_path)
}
\arguments{
\item{rds_files_path}{Path where SummarizedExperiment files live}

\item{output_path}{Path where the plots will ve saved}
}
\description{
Generate RNA QC (PCA, MDS, Sex) plots
}
\author{
Nurlan Kerimov
}
