% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotMDSAnalysis}
\alias{plotMDSAnalysis}
\title{Generate MDS QC Plot for different cell types}
\usage{
plotMDSAnalysis(study_data_se, condition = "all")
}
\arguments{
\item{study_data_se}{SummarizedExperiment file to be analysed}

\item{condition}{Boolean value if html output should be created (Default:FALSE)}
}
\value{
MDS plot of study
}
\description{
Generate MDS QC Plot for different cell types
}
\author{
Nurlan Kerimov
}
