% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotMSDFromMatrix}
\alias{plotMSDFromMatrix}
\title{Generate MDS QC Plot from PCA Matrix}
\usage{
plotMSDFromMatrix(mds_matrix, export_output = FALSE,
  html_output = FALSE, output_dir = "./")
}
\arguments{
\item{mds_matrix}{MDS matrix}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
PCA plot of study
}
\description{
Generate MDS QC Plot from PCA Matrix
}
\author{
Nurlan Kerimov
}
