% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotPCAAnalysis}
\alias{plotPCAAnalysis}
\title{Generate PCA QC Plot for different cell types}
\usage{
plotPCAAnalysis(study_data_se, condition = "all", html_output = FALSE,
  output_dir = "./")
}
\arguments{
\item{study_data_se}{SummarizedExperiment file to be analysed}

\item{condition}{Boolean value if html output should be created (Default:FALSE)}

\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
PCA plot of study
}
\description{
Generate PCA QC Plot for different cell types
}
\author{
Nurlan Kerimov
}
