% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotPCAFromMatrix}
\alias{plotPCAFromMatrix}
\title{Generate PCA QC Plot from PCA Matrix}
\usage{
plotPCAFromMatrix(pca_matrix, export_output = FALSE,
  html_output = FALSE, output_dir = "./")
}
\arguments{
\item{pca_matrix}{PCA matrix}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
PCA plot of study
}
\description{
Generate PCA QC Plot from PCA Matrix
}
\author{
Nurlan Kerimov
}
