% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotSexQC}
\alias{plotSexQC}
\title{Generate Sex dependent QC Plot.}
\usage{
plotSexQC(study_data, html_output = FALSE, output_dir = "./")
}
\arguments{
\item{study_data}{SummarizedExperiment file to be analysed}

\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}
}
\value{
plotly object of SEX QC
}
\description{
Generate Sex dependent QC Plot.
}
\author{
Nurlan Kerimov
}
