% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qc_utils.R
\name{plotSexQCFromMatrix}
\alias{plotSexQCFromMatrix}
\title{Generate Sex dependent QC Plot.}
\usage{
plotSexQCFromMatrix(sex_qc_matrix, study_name, export_output = FALSE,
  html_output = FALSE, output_dir = "./")
}
\arguments{
\item{html_output}{Boolean value if html output should be created (Default:FALSE)}

\item{output_dir}{html file output dir, if html_output is TRUE (Default:current directory)}

\item{study_data}{SummarizedExperiment file to be analysed}
}
\value{
plotly object of SEX QC
}
\description{
Generate Sex dependent QC Plot.
}
\author{
Nurlan Kerimov
}
