% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtltools_normalisation_wrappers.R
\name{qtltoolsPrepareSE}
\alias{qtltoolsPrepareSE}
\title{Normalise phenotype SummarisedExperiment object for QTL mapping with QTLtools}
\usage{
qtltoolsPrepareSE(se, quant_method)
}
\arguments{
\item{se}{SummarizedExperiment object used for QTL mapping}

\item{quant_method}{Quantification method used to generate the data. Valid options are: featureCounts, array, LeafCutter}
}
\value{
Normalised SummarizedExperiment object
}
\description{
Normalise phenotype SummarisedExperiment object for QTL mapping with QTLtools
}
