% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{qtltoolsTabixFetchPhenotypes}
\alias{qtltoolsTabixFetchPhenotypes}
\title{Fetch particular genes from tabix indexed QTLtools output file.}
\usage{
qtltoolsTabixFetchPhenotypes(phenotype_ranges, tabix_file)
}
\arguments{
\item{tabix_file}{Tabix-indexed fastqtl output file.}

\item{gene_ranges}{GRanges object with coordinates of the cis regions around genes.}
}
\value{
List of data frames containing QTLtools results for each gene.
}
\description{
Fetch particular genes from tabix indexed QTLtools output file.
}
