% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{quantileNormaliseCols}
\alias{quantileNormaliseCols}
\title{Perform inverse normal transformation on the columns of a matrix}
\usage{
quantileNormaliseCols(matrix, ...)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix
}
\description{
Perform inverse normal transformation on the columns of a matrix
}
