% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{quantileNormaliseRows}
\alias{quantileNormaliseRows}
\title{Perform inverse normal transformation on the rows of a matrix}
\usage{
quantileNormaliseRows(matrix, ...)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix
}
\description{
Perform inverse normal transformation on the rows of a matrix
}
