% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{quantileNormaliseVector}
\alias{quantileNormaliseVector}
\title{Force a vector of values into standard normal distribution}
\usage{
quantileNormaliseVector(x)
}
\arguments{
\item{x}{numeric vector with arbitrary distribution}
}
\value{
Vector with a standard normal distribution
}
\description{
Force a vector of values into standard normal distribution
}
