% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{reformatPhenotypeId}
\alias{reformatPhenotypeId}
\title{Reformat phenotype IDs according to quantification method}
\usage{
reformatPhenotypeId(read_counts, quant_method = "gene_counts")
}
\arguments{
\item{read_counts}{count matrix of any phenotype quantification.}

\item{quant_method}{Quantification method. Can be gene_counts, leafcutter, txrevise, transcript_usage or exon_counts}
}
\description{
Reformat phenotype IDs according to quantification method
}
\author{
Nurlan Kerimov
}
