% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{replaceNAsWithRowMeans}
\alias{replaceNAsWithRowMeans}
\title{replaceNAsWithRowMeans}
\usage{
replaceNAsWithRowMeans(matrix)
}
\arguments{
\item{matrix}{matrix}
}
\description{
replaceNAsWithRowMeans
}
