% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{studySEtoCountMatrices}
\alias{studySEtoCountMatrices}
\title{Split study SummarizedExperiment object into invidiual input files for QTLTools.}
\usage{
studySEtoCountMatrices(se, assay_name, out_dir, study_name = NULL)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{assay_name}{Name of the assay in the SummarizedExperiment object used for QTL mapping.}

\item{out_dir}{Path to the output directory where the QTLTools input files will be written.}

\item{study_name}{Custom study_name character string to safe files with prefix}
}
\value{
None
}
\description{
Uses the qtl_group metadata column (required) to split a single SummarizedExperiment
object into multiple independent input files for QTLTools. Write the files directly to disk.
}
