% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_QTLTools.R
\name{studySEtoQTLTools}
\alias{studySEtoQTLTools}
\title{Split study SummarizedExperiment object into invidiual input files for QTLTools.}
\usage{
studySEtoQTLTools(se, assay_name, out_dir, extra_qtl_group = NULL)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{assay_name}{Name of the assay in the SummarizedExperiment object used for QTL mapping.}

\item{out_dir}{Path to the output directory where the QTLTools input files will be written.}

\item{extra_qtl_group}{Used to split the datasets using some additional metadata column in additon
to the qtl_group column. This can for example be used to perform sex-specific or population-specific
eQTL analysis.}
}
\value{
None
}
\description{
Uses the qtl_group metadata column (required) to split a single SummarizedExperiment
object into multiple independent input files for QTLTools. Write the files directly to disk.
}
