% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{subsetSEByColumnValue}
\alias{subsetSEByColumnValue}
\title{Extract a subset of the data from a SummarizedExperiment based on column information}
\usage{
subsetSEByColumnValue(se, column, value)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{column}{Name of the column in the colData object}

\item{value}{Value of the column to be extracted}
}
\value{
Subsetted SummmarizedExperiment object
}
\description{
Extract a subset of the data from a SummarizedExperiment based on column information
}
