% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{transformSE_PCA}
\alias{transformSE_PCA}
\title{Quantile normalise SummarizedExperiment by rows}
\usage{
transformSE_PCA(se, assay_name = "cqn", n_pcs = NULL,
  log_transform = FALSE, column_prefix = "",
  feature_id = "sample_id", ...)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{assay_name}{Name of the assay to be processed in the se object}

\item{n_pcs}{Number of desired principal components}

\item{log_transform}{Boolean value}

\item{column_prefix}{custom column_prefix}

\item{feature_id}{default: sample_id}
}
\value{
SummarizedExperiment object with quantile-normalised data in the qnorm assay
}
\description{
Quantile normalise SummarizedExperiment by rows
}
