% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_QC_norm.R
\name{estimateMedianTPM}
\alias{estimateMedianTPM}
\title{Estimate median TPM for each gene in each sample group of a SummarizedExperiment object.}
\usage{
estimateMedianTPM(se, subset_by = "qtl_group", assay_name = "counts")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{subset_by}{Name of the column used for subsetting the data.}

\item{assay_name}{Name of the assay used for TPM normalisation and calculation of median}
}
\value{
data frame of media tpm values per gene and sample group.
}
\description{
Sample groups are defined accroding to the subset_by column of the colData data frame.
}
